Revise NEMO ocean code to ensure reproducibility with land domain exclusion
NEMO ocean core was revised as the reproducibility of model output was not maintained
when switching between different domain decompositions with land exclusion.
In fact, NEMO allows to compute offline the domain decomposition for every combination of PEs
and it is possible to exclude the land-only domain from the computation.
This enables the allocation of PEs only for the active ocean domains saving computational resources.
For example, ORCA1v1 grid with i=19 and j=16 leads to 256 ocean domains and 48 land ones,
by setting in NEMO namelists jpni=19, jpnj=16 and jpnij=256 activates the land exclusion.