Modify st_archive to avoid link and data overlap in case archiving for atm,lnd,rof
The atm,lnd,rof models copy the last output file to the restart folder,
as it is used also for restarting the code.
After the change of restart linking in local execution folder this generated
an overwrite of the previuos last output file in the hist folder.
A chnage in file handling in st_archive solved this issue.